Добавлен функционал, исправлены баги
