+++ react-chat/src/components/ChatList/ChatList.jsx
@@ -1,13 +1,13 @@
 import React, { useState, useEffect } from "react";
-import "./ChatList.css"
+import "./ChatList.css";
 import { loadMessages, saveMessage } from "./../Storage/Storage";
-import { MessageForm, Messages } from "./../Message/Message";
+import { Messages } from "./../Message/Message";
+import { MessageForm } from "./../MessageForm/MessageForm";
 
-const ChatList = () => {
-  const [currentChatId, setCurrentChatId] = useState(null); // Current chat state
-  const [attachedFile, setAttachedFile] = useState(null); // Attached file state
-  const [messages, setMessages] = useState([]); // Messages of current chat
-  const [chats, setChats] = useState([ // Mock chat data
+const ChatList = ({ currentChatId, onOpenChat }) => {
+  const [messages, setMessages] = useState([]); // Сообщения чата
+  const [chats, setChats] = useState([
+    // Моки чатов
     {
       id: 1,
       avatar: "./images/user-icon.svg",
@@ -16,10 +16,10 @@ const ChatList = () => {
       time: "14:23",
       isRead: true,
     },
-    // Add more chat objects here as needed
+    // Добавь больше чатов по необходимости
   ]);
 
-  // Load messages when a new chat is opened
+  // Загрузка сообщений при открытии нового чата
   useEffect(() => {
     if (currentChatId) {
       const loadedMessages = loadMessages(currentChatId);
@@ -27,52 +27,9 @@ const ChatList = () => {
     }
   }, [currentChatId]);
 
-  // Open a chat and set the chat title in the header
-  const openChat = (chatId, chatTitle) => {
-    setCurrentChatId(chatId);
-    document.querySelector(".header_title-text").textContent = chatTitle;
-  };
-
-  // Close the current chat and clear the messages
-  const closeChat = () => {
-    setCurrentChatId(null);
-    setMessages([]);
-  };
-
-  // Handle file selection and validate file size
-  const handleFileChange = (e) => {
-    const file = e.target.files[0];
-    if (file && file.size > 5000000) {
-      alert("Файл слишком большой. Должен быть меньше 5 МБ.");
-      setAttachedFile(null);
-    } else {
-      setAttachedFile(file);
-    }
-  };
-
-  // Handle message form submission, including sending the file if attached
-  const handleSubmit = (e, chatId) => {
-    e.preventDefault();
-    const messageText = e.target["message-text"].value.trim();
-
-    if (messageText || attachedFile) {
-      const newMessage = {
-        text: messageText || "",
-        sender: "Вы",
-        time: new Date().toLocaleTimeString(),
-        file: attachedFile ? URL.createObjectURL(attachedFile) : null,
-      };
-
-      saveMessage(chatId, newMessage);
-      setMessages((prevMessages) => [...prevMessages, newMessage]);
-      setAttachedFile(null);
-      e.target["message-text"].value = ""; // Clear input field
-    }
-  };
-
   return (
     <div className="chat-container">
-      {/* Chat List */}
+      {/* Список чатов */}
       <div
         id="chat-list-component"
         className="chat-list-component"
@@ -88,7 +45,7 @@ const ChatList = () => {
             <button
               key={chat.id}
               className="chat-item"
-              onClick={() => openChat(chat.id, chat.title)}
+              onClick={() => onOpenChat(chat.id, chat.title)}
             >
               <div className="chat-info-wrp">
                 <img src={chat.avatar} alt="Avatar" />
@@ -106,22 +63,13 @@ const ChatList = () => {
         })}
       </div>
 
-      {/* Chat Window */}
+      {/* Окно чата */}
       {currentChatId && (
         <div className="chat-window">
-          <div className="chat-header">
-            <button className="back-button" onClick={closeChat}>
-              <img
-                src="./images/arrow-back.svg"
-                alt="Back to chat list"
-              />
-            </button>
-          </div>
-
-          {/* Messages */}
+          {/* Сообщения */}
           <Messages messages={messages} />
 
-          {/* Message Form */}
+          {/* Форма отправки сообщений */}
           <MessageForm
             chatId={currentChatId}
             onMessageSend={(newMessage) => {
