+++ react-chat/src/components/Header/Header.jsx
@@ -2,52 +2,57 @@ import React, { useState, useEffect } from "react";
 import "./Header.css";
 import Menu from "./../Menu/Menu";
 
-const Header = () => {
+const Header = ({ currentChatTitle, onBackClick }) => {
   const [menuOpen, setMenuOpen] = useState(false);
 
   const toggleMenu = () => {
     setMenuOpen(!menuOpen);
   };
 
-  const closeMenuOnClickOutside = (event) => {
-    const menuButton = document.getElementById("menu-button");
-    const menuContainer = document.getElementById("menu-container");
-    if (
-      !menuButton.contains(event.target) &&
-      !menuContainer.contains(event.target)
-    ) {
-      setMenuOpen(false);
-    }
-  };
-
   useEffect(() => {
-    document.addEventListener("click", closeMenuOnClickOutside);
-    return () => {
-      document.removeEventListener("click", closeMenuOnClickOutside);
-    };
-  }, []);
+    setMenuOpen(false); // Закрываем меню при смене чата
+  }, [currentChatTitle]);
 
   return (
     <header className="header center">
-      {/* Кнопка меню (назад) */}
-      <button className="header_burger-menu">
-        <img src="./images/burger-menu.svg" alt="Кнопка меню" />
-      </button>
+      {/* Кнопка бургер-меню, отображается только в списке чатов */}
+      {!currentChatTitle && (
+        <button className="header_burger-menu" id="menu-button">
+          <img src="./images/burger-menu.svg" alt="Кнопка меню" />
+        </button>
+      )}
 
-      {/* Информация о пользователе */}
+      {/* Название чата или Messenger */}
       <button className="header_title">
-        <div className="header_title-text">Messenger</div>
+        <div className="header_title-text">
+          {currentChatTitle ? currentChatTitle : "Messenger"}
+        </div>
       </button>
 
       {/* Навигационные кнопки */}
       <nav className="header_nav">
+        {/* Кнопка поиска */}
         <button className="header_search">
           <img src="./images/search-icon.svg" alt="Search" />
         </button>
 
-        {/* Кнопка меню (отображается по условию) */}
-        <Menu />
+        {/* Кнопка меню, отображается только в конкретном чате */}
+        {currentChatTitle && (
+          <button className="header_menu-button" onClick={toggleMenu}>
+            <img src="./images/menu-icon.svg" alt="Menu" />
+          </button>
+        )}
+
+        {/* Выпадающее меню */}
+        {menuOpen && <Menu />}
       </nav>
+
+      {/* Кнопка назад при нахождении в чате */}
+      {currentChatTitle && (
+        <button className="back-button" onClick={onBackClick}>
+          <img src="./images/arrow-back.svg" alt="Back to chat list" />
+        </button>
+      )}
     </header>
   );
 };
